requirements.txt
PySide6>=6.5.0
PyYAML>=6.0
